+++ pysui/sui/sui_pgql/pgql_types.py
@@ -101,6 +101,15 @@ class SuiObjectOwnedAddress:
 
 @dataclasses_json.dataclass_json(letter_case=dataclasses_json.LetterCase.CAMEL)
 @dataclasses.dataclass
+class SuiObjectOwnedParent:
+    """Collection of coin data objects."""
+
+    obj_owner_kind: str
+    parent_id: str
+
+
+@dataclasses_json.dataclass_json(letter_case=dataclasses_json.LetterCase.CAMEL)
+@dataclasses.dataclass
 class SuiObjectOwnedImmutable:
     """Collection of coin data objects."""
 
@@ -116,12 +125,12 @@ class SuiCoinObjectGQL(PGQL_Type):
     version: int
     object_digest: str
     balance: str
-    # coin_owner: str
     has_public_transfer: bool
     coin_object_id: str
     object_owner: Union[
         SuiObjectOwnedAddress,
         SuiObjectOwnedShared,
+        SuiObjectOwnedParent,
         SuiObjectOwnedImmutable,
     ]
     previous_transaction: Optional[str] = dataclasses.field(default="")
@@ -145,6 +154,10 @@ class SuiCoinObjectGQL(PGQL_Type):
                 )
             case "Shared":
                 ser_dict["object_owner"] = SuiObjectOwnedShared.from_dict(owner)
+            case "Parent":
+                ser_dict["object_owner"] = SuiObjectOwnedParent(
+                    owner_kind, owner["owner"]["parent_id"]
+                )
             case "Immutable":
                 ser_dict["object_owner"] = SuiObjectOwnedImmutable(owner_kind)
         # ser_dict = ser_dict | res_dict
@@ -197,9 +210,6 @@ class SuiCoinObjectsGQL(PGQL_Type):
         dlist: list[SuiCoinObjectGQL] = []
         for i_coin in in_data["coin_objects"]:
             dlist.append(SuiCoinObjectGQL.from_query(i_coin))
-        # dlist: list[SuiCoinObjectGQL] = [
-        #     SuiCoinObjectGQL.from_query(i_coin) for i_coin in in_data["coin_objects"]
-        # ]
         return SuiCoinObjectsGQL(dlist, ncurs)
 
 
@@ -284,6 +294,7 @@ class ObjectReadGQL(PGQL_Type):
     object_owner: Union[
         SuiObjectOwnedAddress,
         SuiObjectOwnedShared,
+        SuiObjectOwnedParent,
         SuiObjectOwnedImmutable,
     ]
 
@@ -324,6 +335,10 @@ class ObjectReadGQL(PGQL_Type):
                         )
                     case "Shared":
                         res_dict["object_owner"] = SuiObjectOwnedShared.from_dict(owner)
+                    case "Parent":
+                        res_dict["object_owner"] = SuiObjectOwnedParent(
+                            owner_kind, owner["owner"]["parent_id"]
+                        )
                     case "Immutable":
                         res_dict["object_owner"] = SuiObjectOwnedImmutable(owner_kind)
                 # Flatten dictionary
@@ -363,7 +378,6 @@ class EventGQL(PGQL_Type):
 
     package_id: str
     module_name: str
-    # event_type: str
     timestamp: str
     json: str
     sender: Optional[dict]
@@ -376,8 +390,6 @@ class EventGQL(PGQL_Type):
         _fast_flat(in_mod_id, to_merge)
         in_data |= to_merge
         in_data["json"] = json.dumps(in_data["contents"]["json"])
-        # in_data["event_type"] = in_data.pop("type")["event_type"]
-        # in_data["sender"] = [x["address"] for x in in_data["sender"]]
         return EventGQL.from_dict(in_data)